web_micro/marketplace/requirements.txt
flask
grpcio-tools
gunicorn==20.1.0

